+++ src/components/manage/Blocks/Block/BlocksForm.jsx
@@ -30,6 +30,7 @@ const BlocksForm = (props) => {
     metadata,
     manage,
     children,
+    blocksConfig = config.blocks.blocksConfig,
   } = props;
 
   const blockList = getBlocks(properties);
@@ -163,6 +164,7 @@ const BlocksForm = (props) => {
             pathname,
             metadata,
             properties,
+            blocksConfig,
             selected: selectedBlock === childId,
             multiSelected: multiSelected?.includes(childId),
             type: child['@type'],