+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 
 ### Internal
 
+- Add toPublicURL helper @nzambello
+
 ## 12.4.2 (2021-03-29)
 
 ### Bugfix